+++ src/renderer/src/components/right-sidebar/SourceControlTextGenerationDialog.tsx
@@ -1,48 +1,22 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
-import { CheckCircle2, RefreshCw, Save, Sparkles, Terminal, TriangleAlert } from 'lucide-react'
-import { Button } from '@/components/ui/button'
+import React, { useMemo } from 'react'
+import { TriangleAlert } from 'lucide-react'
 import {
   Dialog,
   DialogContent,
   DialogDescription,
-  DialogFooter,
   DialogHeader,
   DialogTitle
 } from '@/components/ui/dialog'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
 import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue
-} from '@/components/ui/select'
-import { AGENT_CATALOG, AgentIcon } from '@/lib/agent-catalog'
-import { planSourceControlTextGeneration } from '@/lib/source-control-generation-plan'
-import {
-  CUSTOM_AGENT_ID,
-  getCommitMessageAgentCapability,
-  isCustomAgentId,
-  listCommitMessageAgentCapabilities
-} from '../../../../shared/commit-message-agent-spec'
-import {
+  normalizeSourceControlAiSettings,
   resolveSourceControlAiForOperation,
   type ResolvedSourceControlAiGenerationParams
 } from '../../../../shared/source-control-ai'
-import {
-  setSourceControlActionDefault,
-  type SourceControlTextActionId
-} from '../../../../shared/source-control-ai-actions'
-import type { SourceControlAiSettings } from '../../../../shared/source-control-ai-types'
-import type { GlobalSettings, Repo, TuiAgent } from '../../../../shared/types'
-import { toast } from 'sonner'
-import { SourceControlActionVariableChips } from '../source-control/SourceControlActionVariableChips'
+import type { SourceControlTextActionId } from '../../../../shared/source-control-ai-actions'
+import type { GlobalSettings, Repo } from '../../../../shared/types'
+import { SourceControlTextGenerationDialogForm } from './SourceControlTextGenerationDialogForm'
 
-type PlanState =
-  | { status: 'idle' }
-  | { status: 'success'; commandLabel: string; delivery: string; caveat: string }
-  | { status: 'error'; error: string }
+export { buildCommitMessageGenerationParams } from './SourceControlTextGenerationParams'
 
 type SourceControlTextGenerationBaseDialogProps = {
   open: boolean
@@ -59,59 +33,6 @@ type SourceControlTextGenerationDialogProps = SourceControlTextGenerationBaseDia
   title: string
   description: string
   generateLabel: string
-}
-
-const UNCONFIGURED_AGENT_SELECT_VALUE = ''
-type CommitMessageGenerationAgentChoice = ResolvedSourceControlAiGenerationParams['agentId'] | ''
-
-function agentLabel(agentId: TuiAgent): string {
-  return AGENT_CATALOG.find((agent) => agent.id === agentId)?.label ?? agentId
-}
-
-export function buildCommitMessageGenerationParams(args: {
-  agentId: CommitMessageGenerationAgentChoice
-  commandTemplate: string
-  agentArgs?: string
-  baseParams: ResolvedSourceControlAiGenerationParams | null
-  settings: Pick<GlobalSettings, 'agentCmdOverrides'> | null | undefined
-}): ResolvedSourceControlAiGenerationParams | null {
-  if (!args.agentId) {
-    return null
-  }
-  if (isCustomAgentId(args.agentId)) {
-    return {
-      agentId: CUSTOM_AGENT_ID,
-      model: '',
-      customPrompt: args.baseParams?.customPrompt,
-      commandInputTemplate: args.commandTemplate,
-      ...(args.agentArgs !== undefined ? { agentArgs: args.agentArgs } : {}),
-      customAgentCommand: args.baseParams?.customAgentCommand ?? ''
-    }
-  }
-  const capability = getCommitMessageAgentCapability(args.agentId)
-  if (!capability) {
-    return null
-  }
-  const sameResolvedAgent = args.baseParams?.agentId === args.agentId
-  const modelId =
-    sameResolvedAgent && args.baseParams?.model
-      ? args.baseParams.model
-      : (capability.models.find((model) => model.id === capability.defaultModelId)?.id ??
-        capability.defaultModelId)
-  const model = capability.models.find((candidate) => candidate.id === modelId)
-  const thinkingLevel =
-    sameResolvedAgent && args.baseParams?.thinkingLevel
-      ? args.baseParams.thinkingLevel
-      : model?.defaultThinkingLevel
-  const agentCommandOverride = args.settings?.agentCmdOverrides?.[args.agentId]?.trim()
-  return {
-    agentId: args.agentId,
-    model: modelId,
-    ...(thinkingLevel ? { thinkingLevel } : {}),
-    commandInputTemplate: args.commandTemplate,
-    ...(args.agentArgs !== undefined ? { agentArgs: args.agentArgs } : {}),
-    ...(agentCommandOverride ? { agentCommandOverride } : {})
-  }
 }
 
 export function SourceControlTextGenerationDialog({
@@ -140,94 +61,22 @@ export function SourceControlTextGenerationDialog({
     [actionId, discoveryHostKey, repo, settings]
   )
   const baseParams = resolved.ok ? resolved.value.params : null
-  const capabilities = useMemo(listCommitMessageAgentCapabilities, [])
-  const showCustomAgent = Boolean(
-    baseParams && (isCustomAgentId(baseParams.agentId) || baseParams.customAgentCommand?.trim())
+  const sourceControlAi = useMemo(
+    () =>
+      settings
+        ? normalizeSourceControlAiSettings(settings.sourceControlAi, settings.commitMessageAi)
+        : null,
+    [settings]
   )
-  const [agentId, setAgentId] = useState<CommitMessageGenerationAgentChoice>('')
-  const [commandTemplate, setCommandTemplate] = useState('{basePrompt}')
-  const [agentArgs, setAgentArgs] = useState('')
-  const [plan, setPlan] = useState<PlanState>({ status: 'idle' })
-  const [saving, setSaving] = useState(false)
-  const commandTemplateId = `source-control-${actionId}-command-template`
-
-  useEffect(() => {
-    if (!open || !baseParams) {
-      return
-    }
-    setAgentId(baseParams.agentId)
-    setCommandTemplate(baseParams.commandInputTemplate ?? '{basePrompt}')
-    setAgentArgs(baseParams.agentArgs ?? '')
-    setPlan({ status: 'idle' })
-  }, [baseParams, open])
-
-  const params = buildCommitMessageGenerationParams({
-    agentId,
-    commandTemplate,
-    agentArgs,
-    baseParams,
-    settings
-  })
-  const paramsPlanResult = params ? planSourceControlTextGeneration(actionId, params) : null
-  const canRunGeneration = Boolean(params && paramsPlanResult?.ok)
-
-  const handlePlan = (): void => {
-    if (!params || !paramsPlanResult) {
-      setPlan({ status: 'error', error: 'Choose an agent before checking generation.' })
-      return
-    }
-    setPlan(
-      paramsPlanResult.ok
-        ? {
-            status: 'success',
-            commandLabel: paramsPlanResult.commandLabel,
-            delivery: paramsPlanResult.delivery,
-            caveat: paramsPlanResult.caveat
-          }
-        : { status: 'error', error: paramsPlanResult.error }
-    )
-  }
-
-  const saveCurrentDefaults = useCallback(
-    async (options: { showToast: boolean; showErrors: boolean }): Promise<boolean> => {
-      if (!params || saving || !paramsPlanResult?.ok) {
-        if (options.showErrors && paramsPlanResult && !paramsPlanResult.ok) {
-          setPlan({ status: 'error', error: paramsPlanResult.error })
-        }
-        return false
-      }
-      setSaving(true)
-      try {
-        await onSaveDefaults(params)
-        if (options.showToast) {
-          toast.success(
-            actionId === 'commitMessage'
-              ? 'Saved commit-message recipe.'
-              : 'Saved hosted-review recipe.'
-          )
-        }
-        return true
-      } finally {
-        setSaving(false)
-      }
-    },
-    [actionId, onSaveDefaults, params, paramsPlanResult, saving]
-  )
-
-  const handleGenerate = (): void => {
-    if (!params || !paramsPlanResult?.ok) {
-      if (paramsPlanResult && !paramsPlanResult.ok) {
-        setPlan({ status: 'error', error: paramsPlanResult.error })
-      }
-      return
-    }
-    onGenerate(params)
-    onOpenChange(false)
-  }
-
-  const handleSaveDefaults = async (): Promise<void> => {
-    await saveCurrentDefaults({ showToast: true, showErrors: true })
-  }
+  const formKey = open
+    ? JSON.stringify([
+        actionId,
+        baseParams?.agentId ?? '',
+        baseParams?.commandInputTemplate ?? '',
+        baseParams?.agentArgs ?? '',
+        baseParams?.customAgentCommand ?? ''
+      ])
+    : 'closed'
 
   return (
     <Dialog open={open} onOpenChange={onOpenChange}>
@@ -237,188 +86,25 @@ export function SourceControlTextGenerationDialog({
           <DialogDescription className="text-xs">{description}</DialogDescription>
         </DialogHeader>
 
-        <div className="space-y-4">
-          {!resolved.ok ? (
-            <p className="flex items-start gap-1.5 rounded-md border border-destructive/30 bg-destructive/5 px-3 py-2 text-xs text-destructive">
-              <TriangleAlert className="mt-px size-3.5 shrink-0" />
-              {resolved.error}
-            </p>
-          ) : null}
+        {!resolved.ok ? (
+          <p className="flex items-start gap-1.5 rounded-md border border-destructive/30 bg-destructive/5 px-3 py-2 text-xs text-destructive">
+            <TriangleAlert className="mt-px size-3.5 shrink-0" />
+            {resolved.error}
+          </p>
+        ) : null}
 
-          <div className="space-y-2">
-            <Label className="text-xs">Agent</Label>
-            <Select
-              value={agentId || UNCONFIGURED_AGENT_SELECT_VALUE}
-              onValueChange={(value) => {
-                if (value === UNCONFIGURED_AGENT_SELECT_VALUE) {
-                  return
-                }
-                setAgentId(value === CUSTOM_AGENT_ID ? CUSTOM_AGENT_ID : (value as TuiAgent))
-                setPlan({ status: 'idle' })
-              }}
-            >
-              <SelectTrigger size="sm" className="h-8 text-xs">
-                <SelectValue placeholder="Choose agent" />
-              </SelectTrigger>
-              <SelectContent>
-                {capabilities.map((capability) => (
-                  <SelectItem key={capability.id} value={capability.id}>
-                    <span className="flex items-center gap-2">
-                      <AgentIcon agent={capability.id} size={14} />
-                      {agentLabel(capability.id)}
-                    </span>
-                  </SelectItem>
-                ))}
-                {showCustomAgent ? (
-                  <SelectItem value={CUSTOM_AGENT_ID}>
-                    <span className="flex items-center gap-2">
-                      <Terminal className="size-3.5 text-muted-foreground" />
-                      Custom command
-                    </span>
-                  </SelectItem>
-                ) : null}
-              </SelectContent>
-            </Select>
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor={`source-control-${actionId}-cli-args`} className="text-xs">
-              CLI arguments
-            </Label>
-            <Input
-              id={`source-control-${actionId}-cli-args`}
-              value={agentArgs}
-              spellCheck={false}
-              placeholder="--model sonnet"
-              onChange={(event) => {
-                setAgentArgs(event.target.value)
-                setPlan({ status: 'idle' })
-              }}
-              className="h-8 font-mono text-xs"
-            />
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor={commandTemplateId} className="text-xs">
-              Command template
-            </Label>
-            <textarea
-              id={commandTemplateId}
-              rows={8}
-              value={commandTemplate}
-              spellCheck={false}
-              onChange={(event) => {
-                setCommandTemplate(event.target.value)
-                setPlan({ status: 'idle' })
-              }}
-              className="w-full resize-y rounded-md border border-border bg-background px-2.5 py-2 font-mono text-xs text-foreground outline-none placeholder:text-muted-foreground/70 focus-visible:ring-1 focus-visible:ring-ring"
-            />
-            <SourceControlActionVariableChips
-              actionId={actionId}
-              onInsert={(variable) => {
-                const separator =
-                  commandTemplate.endsWith('\n') || commandTemplate.length === 0 ? '' : ' '
-                setCommandTemplate(`${commandTemplate}${separator}{${variable}}`)
-                setPlan({ status: 'idle' })
-              }}
-            />
-          </div>
-
-          {plan.status !== 'idle' ? (
-            <div
-              className={
-                plan.status === 'error'
-                  ? 'rounded-md border border-destructive/30 bg-destructive/5 px-3 py-2 text-xs text-destructive'
-                  : 'space-y-1 rounded-md border border-border bg-muted/30 px-3 py-2 text-xs text-muted-foreground'
-              }
-            >
-              {plan.status === 'error' ? (
-                <span className="flex items-start gap-2">
-                  <TriangleAlert className="mt-px size-3.5 shrink-0" />
-                  {plan.error}
-                </span>
-              ) : (
-                <>
-                  <div className="flex items-start gap-2 text-foreground">
-                    <CheckCircle2 className="mt-px size-3.5 shrink-0 text-emerald-500" />
-                    {plan.delivery}
-                  </div>
-                  <div className="truncate font-mono text-[11px]">Launch: {plan.commandLabel}</div>
-                  <div className="text-[11px]">{plan.caveat}</div>
-                </>
-              )}
-            </div>
-          ) : null}
-        </div>
-
-        <DialogFooter className="gap-2">
-          <Button type="button" variant="outline" size="sm" onClick={handlePlan}>
-            <CheckCircle2 className="size-4" />
-            Check generation
-          </Button>
-          <Button
-            type="button"
-            variant="outline"
-            size="sm"
-            disabled={!canRunGeneration || saving}
-            onClick={() => void handleSaveDefaults()}
-          >
-            {saving ? <RefreshCw className="size-4 animate-spin" /> : <Save className="size-4" />}
-            Save as default
-          </Button>
-          <Button
-            type="button"
-            size="sm"
-            disabled={!canRunGeneration || saving}
-            onClick={handleGenerate}
-          >
-            <Sparkles className="size-4" />
-            {generateLabel}
-          </Button>
-        </DialogFooter>
+        <SourceControlTextGenerationDialogForm
+          key={formKey}
+          actionId={actionId}
+          generateLabel={generateLabel}
+          settings={settings}
+          baseParams={baseParams}
+          sourceControlAi={sourceControlAi}
+          onGenerate={onGenerate}
+          onOpenChange={onOpenChange}
+          onSaveDefaults={onSaveDefaults}
+        />
       </DialogContent>
     </Dialog>
   )
-}
-
-export function applySourceControlTextGenerationDefaults(
-  current: SourceControlAiSettings,
-  actionId: SourceControlTextActionId,
-  params: ResolvedSourceControlAiGenerationParams
-): SourceControlAiSettings {
-  if (params.agentId === CUSTOM_AGENT_ID) {
-    const currentRecipe = current.actions?.[actionId] ?? {}
-    const { agentId: _agentId, ...recipeWithoutAgent } = currentRecipe
-    return {
-      ...current,
-      // Why: custom commands are stored as the global Source Control AI
-      // fallback, not as per-action TUI agents; saving must make that
-      // fallback explicit so the action keeps using the custom command.
-      agentId: CUSTOM_AGENT_ID,
-      actions: {
-        ...current.actions,
-        [actionId]: {
-          ...recipeWithoutAgent,
-          commandInputTemplate: params.commandInputTemplate ?? '{basePrompt}',
-          ...(params.agentArgs !== undefined ? { agentArgs: params.agentArgs } : {})
-        }
-      }
-    }
-  }
-  return {
-    ...current,
-    actions: setSourceControlActionDefault(current.actions, actionId, {
-      agentId: params.agentId,
-      commandInputTemplate: params.commandInputTemplate ?? '{basePrompt}',
-      ...(params.agentArgs !== undefined ? { agentArgs: params.agentArgs } : {})
-    })
-  }
-}
-
-export function applyCommitMessageGenerationDefaults(
-  current: SourceControlAiSettings,
-  _hostKey: string,
-  params: ResolvedSourceControlAiGenerationParams
-): SourceControlAiSettings {
-  return applySourceControlTextGenerationDefaults(current, 'commitMessage', params)
 }
