+++ src/renderer/src/components/settings/GitPane.tsx
@@ -1,4 +1,5 @@
 import type { GlobalSettings } from '../../../../shared/types'
+import type { SourceControlAiSettingsPatch } from '../../../../shared/source-control-ai-types'
 import { Input } from '../ui/input'
 import { Label } from '../ui/label'
 import { useAppStore } from '../../store'
@@ -7,20 +8,40 @@ import { SearchableSetting } from './SearchableSetting'
 import { matchesSettingsSearch } from './settings-search'
 import { GitHubRateLimitPanel } from '../github/github-rate-limit-display'
 import { GitLabRateLimitPanel } from '../gitlab/gitlab-rate-limit-display'
+import { AutoRenameBranchFromWorkSetting } from './AutoRenameBranchFromWorkSetting'
+import { AUTO_RENAME_BRANCH_SEARCH_ENTRIES } from './auto-rename-branch-search'
 
 export { GIT_PANE_SEARCH_ENTRIES }
+
+export function shouldShowAutoRenameBranchSetting(
+  searchQuery: string,
+  hasUnsavedBranchPromptChanges: boolean
+): boolean {
+  return (
+    hasUnsavedBranchPromptChanges ||
+    matchesSettingsSearch(searchQuery, AUTO_RENAME_BRANCH_SEARCH_ENTRIES)
+  )
+}
 
 type GitPaneProps = {
   settings: GlobalSettings
   updateSettings: (updates: Partial<GlobalSettings>) => void | Promise<void>
+  writeSourceControlAiSettings: (patch: SourceControlAiSettingsPatch) => Promise<void>
   displayedGitUsername: string
+  hasUnsavedBranchPromptChanges?: boolean
+  onBranchPromptDirtyChange?: (dirty: boolean) => void
+  branchPromptDiscardSignal?: number
   settingsSearchQuery?: string
 }
 
 export function GitPane({
   settings,
   updateSettings,
+  writeSourceControlAiSettings,
   displayedGitUsername,
+  hasUnsavedBranchPromptChanges = false,
+  onBranchPromptDirtyChange,
+  branchPromptDiscardSignal,
   settingsSearchQuery
 }: GitPaneProps): React.JSX.Element {
   const storeSearchQuery = useAppStore((s) => s.settingsSearchQuery)
@@ -140,6 +161,18 @@ export function GitPane({
           />
         </button>
       </SearchableSetting>
+    ) : null,
+    shouldShowAutoRenameBranchSetting(searchQuery, hasUnsavedBranchPromptChanges) ? (
+      <AutoRenameBranchFromWorkSetting
+        key="auto-rename-branch-from-work"
+        settings={settings}
+        updateSettings={updateSettings}
+        writeSourceControlAiSettings={writeSourceControlAiSettings}
+        forceVisible={hasUnsavedBranchPromptChanges}
+        onBranchPromptDirtyChange={onBranchPromptDirtyChange}
+        branchPromptDiscardSignal={branchPromptDiscardSignal}
+        settingsSearchQuery={searchQuery}
+      />
     ) : null,
     matchesSettingsSearch(searchQuery, {
       title: 'GitHub API Budget',
