+++ src/renderer/src/lib/source-control-launch-agent-selection.ts
@@ -1,6 +1,15 @@
 import { AGENT_CATALOG } from '@/lib/agent-catalog'
+import { isCustomAgentId } from '../../../shared/commit-message-agent-spec'
+import type { SourceControlActionRecipe } from '../../../shared/source-control-ai-actions'
 import { filterEnabledTuiAgents } from '../../../shared/tui-agent-selection'
 import type { TuiAgent } from '../../../shared/types'
+
+export function readSourceControlLaunchRecipeAgentId(
+  recipe: Pick<SourceControlActionRecipe, 'agentId'> | null | undefined
+): TuiAgent | null {
+  const agentId = recipe?.agentId
+  return agentId && !isCustomAgentId(agentId) ? agentId : null
+}
 
 export function pickSourceControlLaunchAgent(args: {
   savedAgent?: TuiAgent | null
