+++ src/shared/source-control-ai.test.ts
@@ -1,5 +1,3 @@
-/* eslint-disable max-lines -- Why: source-control AI precedence regressions
-   are easier to audit when the resolver tests stay beside the shared contract. */
 import { describe, expect, it } from 'vitest'
 import { getDefaultSettings } from './constants'
 import {
@@ -9,7 +7,6 @@ import {
   normalizeRepoSourceControlAiOverrides,
   projectSourceControlAiToLegacyCommitMessageAi,
   readSourceControlAiModelChoiceForHost,
-  resolveSourceControlActionRecipe,
   resolveSourceControlAiForOperation,
   resolveSourceControlAiPrCreationDefaults,
   selectSourceControlAiModelChoiceForHost,
@@ -66,36 +63,6 @@ describe('source-control AI resolution', () => {
     expect(resolve('commitMessage').params.model).toBe('gpt-5.5')
     expect(resolve('pullRequest').params.model).toBe('gpt-5.5')
     expect(resolve('branchName').params.model).toBe('gpt-5.5')
-  })
-
-  it('resolves text action CLI args into generation params', () => {
-    const base = settings()
-    base.sourceControlAi = {
-      ...base.sourceControlAi!,
-      actions: {
-        ...base.sourceControlAi!.actions,
-        pullRequest: {
-          commandInputTemplate: '{basePrompt}',
-          agentArgs: '  --model gpt-5.4  '
-        }
-      }
-    }
-
-    expect(
-      resolveSourceControlAiForOperation({
-        settings: base,
-        repo: null,
-        operation: 'pullRequest',
-        discoveryHostKey: 'local'
-      })
-    ).toMatchObject({
-      ok: true,
-      value: {
-        params: {
-          agentArgs: '--model gpt-5.4'
-        }
-      }
-    })
   })
 
   it('resolves PR defaults even when Source Control AI generation is disabled', () => {
@@ -200,56 +167,6 @@ describe('source-control AI resolution', () => {
     })
     expect(result.ok && result.value.params.agentId).toBe('codex')
     expect(result.ok && result.value.params.model).toBe('gpt-5.4')
-  })
-
-  it('treats an explicit action null agent as the default agent, not the global legacy agent', () => {
-    const base = settings()
-    base.defaultTuiAgent = 'codex'
-    base.sourceControlAi = {
-      ...base.sourceControlAi!,
-      agentId: 'claude',
-      selectedModelByAgent: { codex: 'gpt-5.5', claude: 'sonnet' },
-      actions: {
-        ...base.sourceControlAi!.actions,
-        commitMessage: { agentId: null, commandInputTemplate: '{basePrompt}' }
-      }
-    }
-
-    const result = resolveSourceControlAiForOperation({
-      settings: base,
-      repo: null,
-      operation: 'commitMessage',
-      discoveryHostKey: 'local'
-    })
-    expect(result.ok && result.value.params.agentId).toBe('codex')
-    expect(result.ok && result.value.params.model).toBe('gpt-5.5')
-  })
-
-  it('treats a repo action null agent as the default agent, not the global action agent', () => {
-    const base = settings()
-    base.defaultTuiAgent = 'codex'
-    base.sourceControlAi = {
-      ...base.sourceControlAi!,
-      agentId: 'claude',
-      selectedModelByAgent: { codex: 'gpt-5.5', claude: 'sonnet' }
-    }
-    const result = resolveSourceControlAiForOperation({
-      settings: base,
-      repo: {
-        sourceControlAi: {
-          actionOverrides: {
-            commitMessage: {
-              agentId: null,
-              commandInputTemplate: '{basePrompt}'
-            }
-          }
-        }
-      },
-      operation: 'commitMessage',
-      discoveryHostKey: 'local'
-    })
-
-    expect(result.ok && result.value.params.agentId).toBe('codex')
   })
 
   it('lets a global operation model override win over the global default', () => {
@@ -371,230 +288,6 @@ describe('source-control AI resolution', () => {
         prCreationDefaults: { draft: true, openAfterCreate: false }
       }).prCreationDefaults
     ).toMatchObject({ draft: true, openAfterCreate: false })
-  })
-
-  it('resolves launch action recipes from repo overrides over global defaults', () => {
-    const base = settings()
-    base.sourceControlAi = {
-      ...base.sourceControlAi!,
-      actions: {
-        ...base.sourceControlAi!.actions,
-        fixChecks: {
-          agentId: 'claude',
-          commandInputTemplate: '{basePrompt}\n\nglobal',
-          agentArgs: '--model sonnet'
-        }
-      }
-    }
-
-    expect(
-      resolveSourceControlActionRecipe({
-        settings: base,
-        repo: {
-          sourceControlAi: {
-            actionOverrides: {
-              fixChecks: {
-                agentId: 'codex',
-                commandInputTemplate: '  {basePrompt}\n\nrepo  ',
-                agentArgs: '  --model gpt-5.5  '
-              }
-            }
-          }
-        },
-        actionId: 'fixChecks'
-      })
-    ).toEqual({
-      agentId: 'codex',
-      commandInputTemplate: '{basePrompt}\n\nrepo',
-      agentArgs: '--model gpt-5.5'
-    })
-  })
-
-  it('lets repo action recipes explicitly clear inherited CLI args', () => {
-    const base = settings()
-    base.sourceControlAi = {
-      ...base.sourceControlAi!,
-      actions: {
-        ...base.sourceControlAi!.actions,
-        fixChecks: {
-          agentId: 'claude',
-          commandInputTemplate: '{basePrompt}',
-          agentArgs: '--model sonnet'
-        }
-      }
-    }
-
-    expect(
-      resolveSourceControlActionRecipe({
-        settings: base,
-        repo: {
-          sourceControlAi: {
-            actionOverrides: {
-              fixChecks: {
-                agentArgs: null
-              }
-            }
-          }
-        },
-        actionId: 'fixChecks'
-      })
-    ).toEqual({
-      agentId: 'claude',
-      commandInputTemplate: '{basePrompt}',
-      agentArgs: ''
-    })
-  })
-
-  it('preserves existing action templates when legacy settings are only the rollback projection', () => {
-    const source = {
-      ...settings().sourceControlAi!,
-      actions: {
-        ...settings().sourceControlAi!.actions,
-        commitMessage: {
-          agentId: 'codex' as const,
-          commandInputTemplate: 'use $best-commit-msg to write a commit'
-        },
-        branchName: {
-          agentId: 'claude' as const,
-          commandInputTemplate: 'name this branch from {firstPrompt}'
-        }
-      }
-    }
-    const legacy = projectSourceControlAiToLegacyCommitMessageAi(source)
-
-    const merged = mergeLegacyCommitMessageAiIntoSourceControlAi(source, legacy)
-
-    expect(merged.actions?.commitMessage).toEqual({
-      agentId: 'codex',
-      commandInputTemplate: 'use $best-commit-msg to write a commit'
-    })
-    expect(merged.actions?.branchName).toEqual({
-      agentId: 'claude',
-      commandInputTemplate: 'name this branch from {firstPrompt}'
-    })
-  })
-
-  it('does not let rollback enabled changes clobber independent branch action templates', () => {
-    const source = {
-      ...settings().sourceControlAi!,
-      enabled: true,
-      actions: {
-        ...settings().sourceControlAi!.actions,
-        commitMessage: {
-          agentId: 'codex' as const,
-          commandInputTemplate: 'use $best-commit-msg to write a commit'
-        },
-        branchName: {
-          agentId: 'claude' as const,
-          commandInputTemplate: 'name this branch from {firstPrompt}'
-        }
-      }
-    }
-    const legacy = {
-      ...projectSourceControlAiToLegacyCommitMessageAi(source),
-      enabled: false
-    }
-
-    const merged = mergeLegacyCommitMessageAiIntoSourceControlAi(source, legacy)
-
-    expect(merged.enabled).toBe(false)
-    expect(merged.actions?.commitMessage).toEqual({
-      agentId: 'codex',
-      commandInputTemplate: 'use $best-commit-msg to write a commit'
-    })
-    expect(merged.actions?.branchName).toEqual({
-      agentId: 'claude',
-      commandInputTemplate: 'name this branch from {firstPrompt}'
-    })
-  })
-
-  it('does not let stale legacy branch instructions clobber independent branch templates', () => {
-    const source = {
-      ...settings().sourceControlAi!,
-      instructionsByOperation: {
-        ...settings().sourceControlAi!.instructionsByOperation,
-        commitMessage: 'old shared prompt',
-        branchName: 'old shared prompt'
-      },
-      actions: {
-        ...settings().sourceControlAi!.actions,
-        commitMessage: {
-          agentId: 'codex' as const,
-          commandInputTemplate: '{basePrompt}\n\nold shared prompt'
-        },
-        branchName: {
-          agentId: 'claude' as const,
-          commandInputTemplate: 'name this branch from {firstPrompt}'
-        }
-      }
-    }
-    const legacy = {
-      ...projectSourceControlAiToLegacyCommitMessageAi(source),
-      customPrompt: 'rollback changed commit prompt'
-    }
-
-    const merged = mergeLegacyCommitMessageAiIntoSourceControlAi(source, legacy)
-
-    expect(merged.actions?.commitMessage).toEqual({
-      agentId: 'codex',
-      commandInputTemplate: '{basePrompt}\n\nrollback changed commit prompt'
-    })
-    expect(merged.instructionsByOperation.branchName).toBe('old shared prompt')
-    expect(merged.actions?.branchName).toEqual({
-      agentId: 'claude',
-      commandInputTemplate: 'name this branch from {firstPrompt}'
-    })
-  })
-
-  it('lets rollback custom-agent changes clear a commit action agent override', () => {
-    const source = {
-      ...settings().sourceControlAi!,
-      agentId: 'codex' as const,
-      customAgentCommand: '',
-      actions: {
-        ...settings().sourceControlAi!.actions,
-        commitMessage: {
-          agentId: 'codex' as const,
-          commandInputTemplate: 'use $best-commit-msg to write a commit'
-        }
-      }
-    }
-    const legacy = {
-      ...projectSourceControlAiToLegacyCommitMessageAi(source),
-      agentId: 'custom' as const,
-      customAgentCommand: 'custom-agent {prompt}'
-    }
-
-    const merged = mergeLegacyCommitMessageAiIntoSourceControlAi(source, legacy)
-
-    expect(merged.agentId).toBe('custom')
-    expect(merged.customAgentCommand).toBe('custom-agent {prompt}')
-    expect(merged.actions?.commitMessage).toEqual({
-      commandInputTemplate: 'use $best-commit-msg to write a commit'
-    })
-  })
-
-  it('preserves an explicitly empty global text action template as invalid config', () => {
-    const base = settings()
-    base.sourceControlAi = {
-      ...base.sourceControlAi!,
-      actions: {
-        ...base.sourceControlAi!.actions,
-        commitMessage: { commandInputTemplate: '' }
-      }
-    }
-
-    expect(
-      resolveSourceControlAiForOperation({
-        settings: base,
-        repo: null,
-        operation: 'commitMessage',
-        discoveryHostKey: 'local'
-      })
-    ).toEqual({
-      ok: false,
-      error: 'Command template is empty for commit messages.'
-    })
   })
 
   it('maps legacy custom prompt to released split instructions', () => {
